Rename runtime to cells & remove old pod_service
