+++ aurae-client/src/cells/mod.rs
@@ -1,2 +1,1 @@
 pub mod cell_service;
-pub mod pod_service;
