+++ aurae-client/src/lib.rs
@@ -1,9 +1,9 @@
 pub use client::{AuraeClient, AuraeClientError};
 pub use config::{AuraeConfig, AuthConfig, SystemConfig};
 
+pub mod cells;
 mod client;
 mod config;
 pub mod cri;
 pub mod discovery;
 pub mod grpc;
-pub mod runtime;
