+++ aurae-proto/src/lib.rs
@@ -33,6 +33,10 @@
 #![allow(clippy::derive_partial_eq_without_eq)]
 #![allow(clippy::match_single_binding)]
 
+pub mod cells {
+    include!("gen/aurae.cells.v0.rs");
+}
+
 pub mod discovery {
     include!("gen/aurae.discovery.v0.rs");
 }
@@ -49,8 +53,4 @@ pub mod cri {
 
 pub mod observe {
     include!("gen/aurae.observe.v0.rs");
-}
-
-pub mod runtime {
-    include!("gen/aurae.runtime.v0.rs");
 }
