+++ auraed/src/cells/cell_service/cells/cell_name.rs
