+++ auraed/src/cells/cell_service/cells/error.rs
