+++ auraed/src/cells/cell_service/error.rs
