+++ auraed/src/cells/cell_service/executables/executable_name.rs
