--- auraed/src/runtime/pod_service/mod.rs
@@ -1,5 +1,0 @@
-pub use pod_service::PodService;
-
-#[allow(clippy::module_inception)]
-mod pod_service;
-// mod pods;
